+++ src/Api/Models/Counter.cs
@@ -7,6 +7,6 @@ namespace Api.Models
         [BsonId]
         public string Key { get; set; }
 
-        public long Seq { get; set; }
+        public int Seq { get; set; }
     }
 }
